restructure call
