+++ src/components/Activities/components/Calls/components/CallCardsList/components/NoteCard/CallCard.js
