+++ src/components/Activities/components/Calls/components/CallCardsList/components/NoteCard/components/CallBody/CallBody.scss
