+++ src/components/Activities/components/Calls/components/CallCardsList/components/NoteCard/components/CardMaker/CardMaker.js
