+++ src/components/Activities/components/Calls/components/CallCardsList/components/NoteCard/components/CommentBox/CommentBox.scss
