+++ src/components/Activities/components/Calls/components/CallCardsList/components/NoteCard/components/CommentBox/components/CommentCard/CommentCard.scss
@@ -1,4 +1,4 @@
-@import '../../../../../../../../../../../../../style/variable.scss';
+@import '../../../../../../../../../../../../style/variable.scss';
 
 .comment-container {
   background-color: #FEF8F0;
