+++ src/components/Activities/components/Calls/components/CallCardsList/components/NoteCard/components/CommentBox/components/CommentForm/CommentForm.scss
@@ -1,4 +1,4 @@
-@import '../../../../../../../../../../../../../style/variable.scss';
+@import '../../../../../../../../../../../../style/variable.scss';
 
 .comment-form-container {
   font-family: Arial, Helvetica, sans-serif;
