+++ src/components/Activities/components/Calls/components/CallCardsList/components/NoteCard/index.js
