+++ src/components/Activities/components/Calls/components/CallModal/CallModal.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import './Call.scss';
-import CommingSoon from '../../../../../img/Modal/Call/callsImg.jpg';
+import CommingSoon from '../../../../../../img/Modal/Call/callsImg.jpg';
 
 const Calls = () => (
   <div className="iconCotainer">
