+++ src/components/Activities/components/Calls/components/TimeLineControls/TimeLineControls.scss
@@ -1,4 +1,4 @@
-@import '../../../../../../../style/variable.scss';
+@import '../../../../../../style/variable.scss';
 
 .call-timeline-action-container {
   display: flex;
