+++ src/components/Activities/components/Note/components/NoteCardsList/components/NoteCard/components/CreatedBy/CreatedBy.scss
@@ -1,5 +1,6 @@
 .createdby-container {
   display: flex;
+  margin-top: 15px;
 }
 
 .createdby-user-icon {
